spaces/requirements.txt
gradio>=4.20.0
torch
git+https://github.com/openai/CLIP.git
Pillow
numpy
scikit-learn
faiss-cpu
pandas
joblib
